fix(ci): authenticate worker runner onboarding
